Modulariza formularios de la app "core"
